Refactor assignment form
